pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gobox-mcp"
version = "0.1.0"
description = "MCP server for Gobox Open API — 45 tools covering orders, inventory, warehouses, reports, shops, shipments"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Winter279" }]
keywords = ["mcp", "gobox", "wms", "warehouse", "ecommerce", "fastmcp"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "fastmcp>=0.2.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
gobox-mcp = "gobox_mcp.server:main"

[project.urls]
Homepage = "https://github.com/Winter279/gobox-mcp"
Repository = "https://github.com/Winter279/gobox-mcp"
Issues = "https://github.com/Winter279/gobox-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/gobox_mcp"]
